Create email extractor
